+++ src/generate.rs
@@ -432,11 +432,12 @@ pub const fn primes_geq<const N: usize, const MEM: usize>(
 
 /// The error returned by [`primes_lt`] and [`primes_geq`] if the input
 /// is invalid or does not work to produce the requested primes.
-///
-/// Only implements the [`Error`](core::error::Error) trait
-/// if the `std` feature is enabled.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(
+    feature = "rkyv",
+    derive(rkyv::Archive, rkyv::Serialize, rkyv::Deserialize)
+)]
 pub enum GenerationError {
     /// The limit was larger than or equal to `MEM^2`.
     TooSmallSieveSize,