+++ src/sieve.rs
@@ -249,11 +249,12 @@ pub const fn sieve<const N: usize>() -> [bool; N] {
 
 /// The error returned by [`sieve_lt`] and [`sieve_geq`] if the input
 /// is invalid or does not work to sieve the requested range.
-///
-/// Only implements the [`Error`](core::error::Error) trait
-/// if the `std` feature is enabled.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(
+    feature = "rkyv",
+    derive(rkyv::Archive, rkyv::Serialize, rkyv::Deserialize)
+)]
 pub enum SieveError {
     /// The limit was less than or equal to `N` (for `sieve_lt`).
     TooSmallLimit,